requirements.txt
beautifulsoup4>=4.13
requests>=2.32
pillow>=11.2
python-pptx>=1.0.2
